more elegant solution for ifgramStack.datasetNames
adjust to make ifgramStack.datasetNames to be the list of h5py.Dataset name 1) existed in the file regardless of names; 2) with last two dimension equals to length and width; 3) in the order that names in ifgramDatasetNames first and all the others in alphabet order.

move get_poly_mask() and get_circular_mask() from generate_mask.py to utils.py and plot.py for easier usage.

add --sbas option to be the same as --fast option for ifgram_inversion.py

add get_largest_conn_component() to utils.py to extract the largest connected component from an input 2d array

do not write data dataset in writefile.write(), thus only aux dataset are copied to the output file.