Added parallelization using `dask.submit`
This is an initial pass at best. Number of workers is "hardcoded" and
testing is solely done locally. This approach is still a bit slower than
using a single core on a MacBook Pro. Some performance testing and
exploration is needed to see why this is the case.

Next steps:
1. Figure out bug when initializing Client without `processes = False`
2. Test on Pegasus
3. Do some performance testing.