tropo_pyaps: support ${WEATHER_DIR}
add support of identifying environmental variable "WEATHER_DIR" for the default location of downloaded GAMs files.

switch the default value for the weather directory to ${WEATHER_DIR] in both pysarApp.py and tropo_pyaps.py.

Use the current directory "./" as the fallback value if WEATHER_DIR is not found and identified.