remove obsolete utils/plot_map
remove utils/plot_map.py because the utils/plot.draw_scalebar() is only dependent of pyproj and independent of both basemap and cartopy.

Another benefit is that, it allows plotting scalebar all the time without cartopy. This allows the adaptive lat/lon label when zooming and panning, when --lalo-label is not enabled.