plot_smallbaselineApp.sh: running view.py commands in parallel (#345)
Add an '&' to all view.py commands and a 'wait' after the last one, so that they all run in parallel. 

On stampede this reduces the runtime from 2 minutes to 10 seconds. On systems that have less cores than commands it is supposed to run as much commands in parallel as cores are available. This worked for me on linux (16 core machine, pegasus login node) and macOS (6 cores).