prep_hyp3: change LAT/LON_REF1/2/3/4 for ascending geometries
The LAT/LON_REF1/2/3/4 attributes are in a different order for
ascending and descending geometries. This commit reworks the scirpt so
that this difference can be accounted for. The dem file origionally did not
have an external metadata file, so adding in this functionally created
a large changes in the script. Also updated docs/dir_structure.md to
reflect these changes.

See Pull Request: #542