spatial_filter: add double difference filter option
The double difference filter as used in Bekaert et al 2020
has been shown to effectively filter out non-local modes of
deformation. This particularly when using MintPy to monitor
local forms of deformation, such as landslides. This commit
implements the double difference filter in spatial_filter.py
and reformats the filter_par parameter so that it accepts
multiple arguments.