spatial_filter: add double difference filter option (#552)
* spatial_filter: add `double_difference` filter option

The double difference filter as used in Bekaert et al. (2020, RSE) has been shown to effectively filter out non-local modes of deformation. This particularly when using MintPy to monitor local forms of deformation, such as landslides. 

This commit implements the double_difference filter in spatial_filter.py and reformats the filter_par parameter so that it accepts multiple arguments, as required by double_difference.

This PR also fixes missing the missing dataset bug when datasets were specified, ONLY the datasets specified
were written to the output file. This patch ensures that all datasets present in the input file are present in the output file, even if they are unfiltered.