Proposed changes on this code to use the attribute from timeseries file instead of the geometry file
I've proposed changes on this code to used the attributes from the timeseries file instead of the geometry file. This implementation is made due to the possbile incomplete geometryRadar.h5 attributes on the alosStack. This edit is also consistent with the workflow used by tropo_pyaps3.py which uses the timeseries file instead of the geometry file for the attributes.