solid_earth_tides: bugfix of missing CENTER_LINE_UTC (#568)
I've proposed to use the attributes from the time-series file instead of the geometry file. This implementation is made due to the possible incomplete geometryRadar.h5 attributes on the alosStack. This edit is also consistent with the workflow used by tropo_pyaps3.py which uses the time-series file instead of the geometry file for the attributes.