Properly pass the azimuth variable
The azimuth variable is supplied as part of the command line input, however it's not actually passed to the `get_design_matrix` function where its needed, meaning the command line variable does nothing and the azimuth angle defaults to 90 degrees. I've added code to pass it through to the design matrix function, so now the azimuth angle can be modified via the command line.