improved auto lat/lon step for geocoding
Add utils0.auto_lat_lon_step_size() to calculate the default output lat/lon steps for geocoding.

By default, the current version ensure the same number of rows and columns between input and output files; while this new version ensures the same pixel area between the input and output files and with proper dimension ratio using the bounding box of an rotated rectangle. Thus, this is a better version, especially when the azimuth angle is far from the pure north-south, such as the UAVSAR flight path, or when the space-borne SAR track is long.