`timeseries2velocity`: add `--polyline` option (#999)
This PR adds the `timeseries2velocity.py --polyline` option (and for `tsview.py` as well). Given one or multiple dates, this option will estimate an additional linear slope starting from the specified date, to catch the linear velocity change of a time series. The result is saved in the velocity.h5 file as a dataset named `velocityPost{YYYYMMDD}`, which is the linear velocity during this segment time period, i.e. the sum of the velocity for the entire period (from the regular polynomial fitting), plus all the additional velocity change before & during the current segment periods.

Detailed changes are as below:

+ `utils.time_func`: construct a design matrix for polyline fitting with the manually specified starting date for each additional line segment

+ `utils.arg_utils.add_timefunc_argument()`: add `--polyline` option

+ `timeseries2velocity`:
   - `model2hdf5_dataset()`: support the polyline function and save the total/cumulative velocity for each polyline segment for easy visualization and interpretation.
   - check input `--polynomial` and `--polyline` options

+ `cli.timeseries2velocity`: add example usage

Other changes not relevant to `--polyline`:

+ `utils.ptime`:
   - `date_list2tbase()`: add `ref_date` arg
   - `date_list2vector()`: add `seconds` arg

+ `utils.arg_utils.add_figure_argument()`: add `--cbar-ticks` option for accurate control on the colorbar tick positions.

+ `simulation.iono.lalo_ground2iono()`: more comments

+ `tsview`: printout msg minor adjustment
   - remove redundant print cmd. Now the tsview cmd msg will always be printed out.
   - always printout the save fig/txt file info, as it's very informative

+ `objects.gps`: change the default folder to store the site list file from `./` to `./GPS`.